+++ crates/full-node/sov-stf-runner/src/lib.rs
@@ -12,7 +12,7 @@ mod state_manager;
 
 pub use crate::config::{
     from_toml_path, CorsConfiguration, HttpServerConfig, MonitoringConfig, ProofManagerConfig,
-    RollupConfig, RunnerConfig, StorageConfig, TelegrafSocketConfig,
+    RollupConfig, RunnerConfig, TelegrafSocketConfig,
 };
 pub use crate::http::rpc_module_to_router;
 pub use crate::runner::*;